cppcheck warnings corrected

+++ src/phase/rtrace.c
@@ -1,6 +1,6 @@
 /*   File      : /afs/psi.ch/user/f/flechsig/phase/src/phase/rtrace.c */
 /*   Date      : <23 Mar 04 11:27:42 flechsig>  */
-/*   Time-stamp: <13 May 13 14:20:31 flechsig>  */
+/*   Time-stamp: <2014-02-16 19:49:37 flechsig>  */
 /*   Author    : Uwe Flechsig, flechsig@psi.ch */
 
 /*   $Source$  */
@@ -480,7 +480,7 @@ void RayTraceSingleRay(struct BeamlineType *bl)
 	else 
 	  printf("RayTraceSingleRay:   Source to image calculation.\n");
 
-	printf("RayTraceSingleRay:   Beamline contains %d element(s).\n", 
+	printf("RayTraceSingleRay:   Beamline contains %u element(s).\n", 
 	       bl->elementzahl);
 
 	while (elcounter< bl->elementzahl)
@@ -963,7 +963,7 @@ void ReAllocResult(struct BeamlineType *bl, int newtype, int dim1, int dim2)
     }
   bl->RESULT.typ= newtype;
 #ifdef DEBUG1 
-  printf("\n&&&&&&&&&&&&&&&&&debug %s AllocResult %d\n", __FILE__, bl->RESULT.RESp);
+  printf("\n&&&&&&&&&&&&&&&&&debug %s AllocResult %d\n", __FILE__, (int)bl->RESULT.RESp);
 #endif
 } /* AllocResult */
 
@@ -974,7 +974,7 @@ void FreeResultMem(struct RESULTType *Re)
   struct PSDType *PSDp;
 
 #ifdef DEBUG1  
-  printf("\n&&&&&&&&&&&&&&&&& debug:  %s FreeResultMem %d\n", __FILE__, Re->RESp); 
+  printf("\n&&&&&&&&&&&&&&&&& debug:  %s FreeResultMem %d\n", __FILE__, (int)Re->RESp); 
 #endif
   
   if (Re->RESp == NULL)
